cataract_backend/cataract_detection/requirements.txt
torch
torchvision
timm
numpy
Pillow
gradio
scikit-learn

fastapi
uvicorn[standard]
python-multipart